Plugin Directory: PCP Scanner: Cast active installs to integer to avoid PHP Fatal errors in `number_format_i18n()` calling `number_format()` with an empty string.
Fixes fatals:
> E_ERROR: Uncaught TypeError: number_format(): Argument #1 ($num) must be of type float, string given


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@14635 74240141-8908-4e6f-9713-ba540dce6ec7

+++ wordpress.org/public_html/wp-content/plugins/plugin-directory/jobs/class-plugin-updates-pcp.php
@@ -240,7 +240,7 @@ class Plugin_Updates_PCP {
 			return;
 		}
 
-		$active_installs = get_post_meta( $plugin->ID, 'active_installs', true );
+		$active_installs = (int) get_post_meta( $plugin->ID, 'active_installs', true );
 
 		$body = sprintf( "Detected errors in *%s*\n", $plugin->post_title );
 		if ( $active_installs >= 10000 ) {
